+++ README.md
@@ -9,8 +9,8 @@
 	- [x] [Llama3 Family](models/llama3/)
 	- [x] [Mistral 7B](models/mistral_rolling/)
 	- [x] [Mixtral 8x7B](models/mixtral.py)
+	- [x] [Qwen QwQ 32B](models/qwq/)
 	- [ ] DeepSeek ...
-	- [ ] Qwen ...
 	- [ ] Gemini ...
 	- [ ] Grok
